+++ src/Resident.h
@@ -10,5 +10,7 @@
 #include "ResidentLuaModule.h"
 #include "ResidentStatusLED.h"
 #include "ResidentStatusDisplay.h"
+#include "ResidentSystemLED.h"
+#include "ResidentSystemDisplay.h"
 
 #endif // RESIDENT_H